pyproject.toml
[tool.poetry]
name = "biostar"
version = "0.1.0"
description = "Bioburden and Sampling Optimization Tool for Assessing Risk (BioSTAR)"
authors = ["Patrick Bjornstad <patrick.t.bjornstad@jpl.nasa.gov>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
python-dotenv = "^1.0"
numpy = "^1.26"
scipy = "^1.10"
pandas = "^2.1"
openpyxl = "^3.1.5"
xlsxwriter = "^3.2.5"
plotly = "~5"
dash = "^3.0"
dash-mantine-components = "^1.1.0"
dash-iconify = "^0.1.2"
deepdiff = "^8.4.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
extend-select = ["D101", "D102", "D103", "I001"]

[tool.ruff.lint.mccabe]
max-complexity = 8

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
